requirements.txt
async-timeout==4.0.3
asyncio==3.4.3
azure-ai-formrecognizer==3.2.1
azure-cognitiveservices-speech==1.32.1
azure-common==1.1.28
azure-core==1.29.4
azure-core-tracing-opentelemetry
opentelemetry-sdk
azure-cosmos==4.5.1
azure-identity==1.14.0
azure-search-documents==11.4.0b8
beautifulsoup4==4.12.2
docx2txt==0.8
evaluate==0.4.0
faiss-cpu==1.7.4
fastapi==0.104.1
fuzzywuzzy==0.18.0
html2text==2020.1.16
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.17.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
langchain==0.0.327
langchain-experimental==0.0.16
langcodes==3.3.0
langsmith==0.0.54
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
multidict==6.0.4
num2words==0.5.12
numexpr==2.8.5
numpy==1.25.2
openai==0.27.2
openpyxl==3.1.2
pandas==2.1.0
prettytable==3.9.0
pydantic_core==2.6.3
pydantic==2.4.2
pydantic-settings==2.0.3
pydeck==0.8.1b0
pyee==9.0.4
Pygments==2.15.1
PyJWT==2.8.0
pyodbc==4.0.39
pyparsing==3.1.1
pypdf==3.16.1
PyPDF3==1.0.6
python-dotenv==1.0.0
python-Levenshtein==0.23.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.2
SQLAlchemy==1.4.47
statsmodels==0.14.0
streamlit==1.27.2
tenacity==8.2.3
tiktoken==0.4.0
tqdm==4.65.0
uvicorn==0.23.0
validators==0.20.0
wget==3.2